% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrix.R
\name{preprocess_frag_bed}
\alias{preprocess_frag_bed}
\title{Preprocess the fragment BED-like data frame}
\usage{
preprocess_frag_bed(frag, bin_size)
}
\value{
A list containing fragment lengths
}
\description{
This function takes a fragment BED as input. For each chromosome, return a
new data frame where each row is for a specific position, associated with a
vector that represents the lengths of all fragments at that position. The
result is much more compact than the original BED-like data frame.
}
\details{
Example of returned value:
$\code{21}
}
\section{A tibble: 74 x 2}{
start len
<int> <list>
1  9000000 <int \link{5,443}>
2  9500000 <int \link{59,183}>
3 10000000 <int \link{45,040}>
4 10500000 <int \link{96,504}>
}

