% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrix.R
\name{read_fragments}
\alias{read_fragments}
\title{Read a fragment data file}
\usage{
read_fragments(file_path, range = NULL, genome = NULL)
}
\arguments{
\item{file_path}{a character vector representing paths of fragment files.
This argument accepts multiple input files which is convenient for sample
pooling.}

\item{range}{read the entire file if \code{range} is \code{NULL}.}
}
\description{
A fragment data file is essentially just a BED file. The reason we need a
special function here to load a fragment data file is that, by convention,
the data frame should contain a column representing MAPQ scores. However, the
data file may not have a column header. Therefore, we may need to guess which
column represents MAPQ scores.
}
