% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgEncodeDacMapabilityConsensusExcludable.hs37-1kg}
\alias{wgEncodeDacMapabilityConsensusExcludable.hs37-1kg}
\title{ENCODE black list}
\format{
An object of class \code{GRanges} of length 411.
}
\usage{
wgEncodeDacMapabilityConsensusExcludable.hs37-1kg
}
\description{
ENCODE black list
}
\keyword{datasets}
